[refactor]ディレクトリ整理
