+++ client/src/components/page/space/SpacePageMain.tsx
@@ -6,7 +6,7 @@ import { MemberBlock } from "./memberBlock/MemberBlock";
 import { TimetableBlock } from "./timetableBlock/TimetableBlock";
 import { ConfigBlock } from "./configBlock/ConfigBlock";
 import { CommentBlock } from "./commentBlock/CommentBlock";
-import { ReactionBlock } from "./ReactionBlock";
+import { ReactionBlock } from "./reactionBlock/ReactionBlock";
 
 import {
   RoomResponse,
